Move polyswarmd method-patching into `tests/conftest.py`
